+++ src/pmdas/linux_zfs/zfs_vdev_mirrorstats.h
@@ -1,11 +1,11 @@
 typedef struct zfs_vdev_mirrorstats {
-    unsigned int rotating_linear;
-    unsigned int rotating_offset;
-    unsigned int rotating_seek;
-    unsigned int non_rotating_linear;
-    unsigned int non_rotating_seek;
-    unsigned int preferred_found;
-    unsigned int preferred_not_found;
+    uint64_t rotating_linear;
+    uint64_t rotating_offset;
+    uint64_t rotating_seek;
+    uint64_t non_rotating_linear;
+    uint64_t non_rotating_seek;
+    uint64_t preferred_found;
+    uint64_t preferred_not_found;
 } zfs_vdev_mirrorstats_t;
 
 void zfs_vdev_mirrorstats_refresh(zfs_vdev_mirrorstats_t *vdev_mirrorstats, regex_t *rgx_row);
